+++ src/helpers/helpers.js
@@ -117,23 +117,6 @@ export const integrateData = (remoteWishList, localWishList) => {
   setStorageData('wishList', localWishList);
 };
 
-export const integrateCart = (remoteCart) => {
-  const localCart = getStorageData('shoppingCart');
-  if (localCart) {
-    remoteCart.forEach(remoteItem => {
-      if (!localCart.find(localItem => localItem.productId === remoteItem.productId)) {
-        localCart.push(remoteItem);
-      }
-    });
-    localCart.forEach(localItem => {
-      if (!remoteCart.find(remoteItem => remoteItem.productId === localItem.productId)) {
-        cartHandler([localItem]);
-      }
-    });
-  }
-  setStorageData('shoppingCart', localCart);
-};
-
 export const toggleWishItems = (productId) => {
   const userId = getUserIdFromCookie();
   const wishListLocal = getStorageData('wishList');
@@ -164,82 +147,6 @@ export const toggleWishItems = (productId) => {
         });
     }
   }
-};
-
-const cartHandler = (products) => {
-  const userId = getUserIdFromCookie();
-  const cartId = getStorageData('cartId');
-  if (userId) {
-    AjaxUtils.ShopCart.getUserShopCart(userId)
-      .then(result => {
-        if (result.message) {
-          AjaxUtils.ShopCart.createShopCart(userId, products)
-            .then(result => {
-              // todo nice popup
-              if (result && result._id) {
-                setStorageData('cartId', result._id);
-              }
-            });
-        } else {
-          AjaxUtils.ShopCart.updateShopCartById(result._id, products, result.userId)
-            .then(result => {
-              // todo nice popup
-              if (result && result._id) {
-                setStorageData('cartId', result._id);
-              }
-            });
-        }
-      });
-  } else if (!userId && cartId.length) {
-    AjaxUtils.ShopCart.updateShopCartById(cartId, products)
-      .then(result => {
-        // todo nice popup
-        console.log('updating for unregistered user', result);
-      });
-  } else {
-    AjaxUtils.ShopCart.createShopCart(null, products)
-      .then(result => {
-        // todo nice popup
-        console.log(result);
-        if (result && result.cart) setStorageData('cartId', result.cart._id);
-      });
-  }
-};
-
-export const addToCart = (productId, cartQuantity = 1, sizeId = '') => {
-  const shopCartLocal = getStorageData('shoppingCart');
-  const product = {
-    productId,
-    cartQuantity,
-    sizeId
-  };
-
-  const itemInCart = shopCartLocal.find(item => item.productId === productId);
-  if (itemInCart) {
-    if (sizeId !== itemInCart.sizeId) {
-      const newItem = {
-        productId,
-        cartQuantity,
-        sizeId
-      };
-      setStorageData('shoppingCart', [...shopCartLocal, newItem]);
-      return;
-    }
-    if (cartQuantity !== itemInCart.cartQuantity) {
-      setStorageData('shoppingCart', [...shopCartLocal, itemInCart.cartQuantity = cartQuantity]);
-    }
-  } else {
-    setStorageData('shoppingCart', [...shopCartLocal, product]);
-  }
-
-  cartHandler(getStorageData('shoppingCart'));
-};
-
-export const deleteFromCart = (productId) => {
-  const shopCartLocal = getStorageData('shoppingCart');
-  const products = shopCartLocal.filter(item => item.productId !== productId);
-  setStorageData('shoppingCart', products);
-  cartHandler(products);
 };
 
 export const defineSortData = (option) => {
